moved payment method + added endpoints
