+++ backend/app.py
@@ -4,7 +4,7 @@ from config import ApplicationConfig
 from flask_bcrypt import Bcrypt
 from flask_cors import CORS
 from flask_session import Session
-from endpoints.auth import auth, bcrypt
+from endpoints.auth import auth, bcrypt, payment
 
 app = Flask(__name__)
 app.config.from_object(ApplicationConfig)
@@ -20,6 +20,7 @@ with app.app_context():
     
  
 app.register_blueprint(auth)
+app.register_blueprint(payment)
 
 if __name__ == "__main__":
     app.run(debug=True)
