+++ backend/endpoints/__init__.py
@@ -1,1 +1,2 @@
 from .auth import auth
+from .payments import payment
